Create Repeated_model _testing
this is what the example was asking for :"a script that allows you to test a linear regression model multiple times. Place the script in a separate function that will take as an argument the number of repetitions that need to be performed. Hint: use a for loop in your function. Each time the data set is to be divided into subsets: training and testing in a random way, for this purpose do not provide the random_state argument so that the results are different each time. The average value of the mean_absolute_percentage_error measure is to be returned as the result of the experiment - i.e. the average percentage of the regression error" 
I also added code to create plots showing the results of the first test as well as of the last test of the loop to show the progress the model made.