Create Outliers.py
the task:"Modify Task 2.2 by adding the outlier removal/replacement procedure. Compare your results in the previous task with your new results."
I made 3 functions, one removing outliers, one replacing them with the mean value of the y_train and one, which does not modify the outlier values. Each of the functions returns the mean_absolute_percentage_error which allows me to later print and compare the results for each of the functions